requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
python-dotenv
pandas>=1.3.0
numpy>=1.21.0
openai>=1.12.0
tweepy>=4.14.0
supabase>=2.3.1
pytz